refactor(api): separate legacy calls
