+++ src/common/api/_legacy/removeItem.js
