+++ src/containers/collections/collections.state.js
@@ -2,9 +2,9 @@ import { select, takeLatest, put } from 'redux-saga/effects'
 import { deriveCollection } from 'common/api/derivers/item'
 import { getCollectionBySlug } from 'common/api/collections'
 import { getCollections } from 'common/api/collections'
-import { saveItem } from 'common/api/saveItem'
-import { removeItemByUrl } from 'common/api/removeItem'
-import { saveItems } from 'common/api/saveItem'
+import { saveItem } from 'common/api/_legacy/saveItem'
+import { removeItemByUrl } from 'common/api/_legacy/removeItem'
+import { saveItems } from 'common/api/_legacy/saveItem'
 
 import { HYDRATE } from 'actions'
 import { COLLECTIONS_HYDRATE } from 'actions'
