+++ src/containers/syndicated-article/syndicated-article.state.js
@@ -9,8 +9,8 @@ import {
   ARTICLE_UNSAVE_FAILURE
 } from 'actions'
 import { getSyndicatedArticle, getRandomSyndicatedArticle } from 'common/api/syndicated-article'
-import { saveItem } from 'common/api/saveItem'
-import { removeItem } from 'common/api/removeItem'
+import { saveItem } from 'common/api/_legacy/saveItem'
+import { removeItem } from 'common/api/_legacy/removeItem'
 import { HYDRATE } from 'actions'
 
 /** ACTIONS
