+++ src/containers/topic/topic.state.js
@@ -1,7 +1,5 @@
 import { put, takeEvery } from 'redux-saga/effects'
-import { getNewTopicFeed } from 'common/api/topics'
-import { deriveRecommendation } from 'common/api/derivers/item'
-import { arrayToObject } from 'common/utilities'
+import { getTopicFeed } from 'common/api/topics'
 
 import { TOPIC_HYDRATE } from 'actions'
 import { TOPIC_SAVE_REQUEST } from 'actions'
@@ -84,7 +82,7 @@ export async function fetchTopicData(topic) {
     // 100% curated, so we need to ask for more `curated` and omit the
     // algorithmic results
 
-    const { itemsById, curatedItems, algorithmicItems } = await getNewTopicFeed(topic, 30)
+    const { itemsById, curatedItems, algorithmicItems } = await getTopicFeed(topic, 30)
 
     return {
       topic,
